Restructure project with modular architecture
