+++ index.html
@@ -76,6 +76,6 @@
                     <img src='//cdn.weatherapi.com/v4/images/weatherapi_logo.png' alt="Weather data by WeatherAPI.com" border="0">
                 </a>
         </footer>
-    <script src="js/components.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
