+++ package.json
@@ -1,0 +1,11 @@
+{
+  "name": "climapp",
+  "version": "1.0.0",
+  "description": "Simple weather app",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js"
+  },
+  "author": "",
+  "license": "ISC"
+}
