+++ server.js
@@ -1,0 +1,39 @@
+const http = require('http');
+const fs = require('fs');
+const path = require('path');
+
+const port = process.env.PORT || 3000;
+
+http.createServer((req, res) => {
+  let filePath = '.' + req.url;
+  if (filePath === './') filePath = './index.html';
+  const ext = path.extname(filePath);
+  let contentType = 'text/html';
+
+  switch (ext) {
+    case '.js':
+      contentType = 'text/javascript';
+      break;
+    case '.css':
+      contentType = 'text/css';
+      break;
+    case '.png':
+      contentType = 'image/png';
+      break;
+    case '.json':
+      contentType = 'application/json';
+      break;
+  }
+
+  fs.readFile(filePath, (err, content) => {
+    if (err) {
+      res.writeHead(404);
+      res.end('File not found');
+    } else {
+      res.writeHead(200, { 'Content-Type': contentType });
+      res.end(content, 'utf-8');
+    }
+  });
+}).listen(port, () => {
+  console.log(`Server running at http://localhost:${port}`);
+});
